Implemented enqueue and dequeue methods for TrafficList and also made Vehicle a public class
